requirements.txt
absl-py
tflearn
tqdm
opt_einsum
keras_preprocessing
argon2-cffi
astunparse
async-generator
attrs
backcall
bleach
cachetools
certifi
cffi
chardet
click
colorama
cycler
decorator
defusedxml
entrypoints
Flask
Flask-SocketIO
Flask-WTF
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
idna
importlib-metadata
imutils
ipykernel
ipython-genutils
itsdangerous
jedi
Jinja2
joblib
jsonschema
Keras
Keras-Preprocessing
kiwisolver
Markdown
matplotlib
mistune
mtcnn
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
oauthlib
opencv-contrib-python-headless
opt-einsum
packaging
pandas
pandocfilters
parso
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pycparser
Pygments
pyparsing
pyrsistent
python-dateutil
python-engineio
python-socketio
pytz
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
requests
requests-oauthlib
rsa
scikit-learn
scipy
Send2Trash
six
sklearn
statsmodels
tabula-py
tensorboard
tensorboard-plugin-wit
tensorflow-cpu
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
tornado
urllib3
waitress
wcwidth
webencodings
Werkzeug
widgetsnbextension
wincertstore
wrapt
WTForms
XlsxWriter
zipp
opencv-python
scipy
python-osc
SQLAlchemy
Flask-SQLAlchemy